+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.4] - 2025-11-28
+
+### Fixed
+
+*   **Changelog:** Corrected changelog history for versions 0.3.2 and 0.3.3.
+
+## [0.3.3] - 2025-11-28
+
+### Added
+
+*   **Feature:** Added retry logic to `AsyncExecutor.map` and `AsyncExecutor.submit`.
+    *   `submit` now accepts `retries` and `retry_delay` arguments.
+    *   `map` now accepts `retries` and `retry_delay` arguments.
+
 ## [0.3.2] - 2025-11-28
 
 ### Added